requirements.txt
discord-py-interactions
python-dotenv
httpx[http2]
yapf